+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/datagen/tag/ModItemTagsProvider.kt
@@ -298,8 +298,6 @@ class ModItemTagsProvider(
 			)
 			.add(
 				ModItems.BREADTH_FIRST_SEARCH_UPGRADE,
-				ModItems.DEPTH_FIRST_SEARCH_UPGRADE,
-				ModItems.PSEUDO_ROUND_ROBIN_UPGRADE,
 				ModItems.ENDER_TRANSMITTER,
 				ModItems.ENDER_RECEIVER,
 				ModItems.PRECISION_SHEARS,